Seperating intermediate escalation, compensation and error tests into their own class
