+++ jbpm-bpmn2/src/test/resources/escalation/BPMN2-EscalationBoundaryEvent.bpmn2
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?> 
+<definitions id="Definition"
+             targetNamespace="http://www.example.org/MinimalExample"
+             typeLanguage="http://www.java.com/javaTypes"
+             expressionLanguage="http://www.mvel.org/2.0"
+             xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
+             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+             xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL BPMN20.xsd"
+             xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI"
+             xmlns:dc="http://www.omg.org/spec/DD/20100524/DC"
+             xmlns:di="http://www.omg.org/spec/DD/20100524/DI"
+             xmlns:tns="http://www.jboss.org/drools">
+
+  <escalation id="MyEscalation" escalationCode="MyEscalation" />
+  <process processType="Private" isExecutable="true" id="EscalationBoundaryEvent" name="Escalation Boundary Event Process" >
+
+    <startEvent id="_1" name="StartProcess" />
+    <sequenceFlow id="_1-_2" sourceRef="_1" targetRef="_2" />
+    
+    <subProcess id="_2" name="Hello"  >
+      <startEvent id="_2-1" name="StartSubProcess" />
+      <sequenceFlow id="_2-1-_2-2" sourceRef="_2-1" targetRef="_2-2" />
+      
+      <endEvent id="_2-2" name="EscalationEvent" >
+        <escalationEventDefinition escalationRef="MyEscalation" />
+      </endEvent>
+    </subProcess>
+    
+    <endEvent id="_3" name="EndProcess" >
+        <terminateEventDefinition/>
+    </endEvent>
+    <sequenceFlow id="_2-_3" sourceRef="_2" targetRef="_3" />
+    
+    <boundaryEvent id="_4" name="EscalationEvent" attachedToRef="_2" cancelActivity="false" >
+      <escalationEventDefinition escalationRef="MyEscalation" />
+    </boundaryEvent>
+    <sequenceFlow id="_4-_5" sourceRef="_4" targetRef="_5" />
+    
+    <scriptTask id="_5" name="Goodbye" >
+      <script>System.out.println("Escalation handled");</script>
+    </scriptTask>
+    <sequenceFlow id="_5-_6" sourceRef="_5" targetRef="_6" />
+    
+    <endEvent id="_6" name="EndProcess" >
+        <terminateEventDefinition/>
+    </endEvent>
+  </process>
+
+</definitions>
