+++ jbpm-bpmn2/src/test/resources/escalation/BPMN2-EscalationEndEvent.bpmn2
