+++ jbpm-bpmn2/src/test/resources/escalation/BPMN2-IntermediateThrowEventEscalation.bpmn2
